requirements.txt
discord.py
pyautogui
psutil
requests
sounddevice
scipy
pywin32
pyinstaller
colorama